API: Move CLI modules into commandline package.
I have seen 'cli' confuse alpha-users who think it means 'client'.
