+++ setup.py
@@ -29,6 +29,7 @@ setup(name='caproto',
                   'protocol',
       packages=['caproto',
                 'caproto.benchmarking',
+                'caproto.commandline',
                 'caproto.curio',
                 'caproto.examples',
                 'caproto.ioc_examples',
@@ -40,10 +41,10 @@ setup(name='caproto',
                 ],
       entry_points={
           'console_scripts': [
-              'caproto-get = caproto.get:main',
-              'caproto-put = caproto.put:main',
-              'caproto-monitor = caproto.monitor:main',
-              'caproto-repeater = caproto.repeater:main',
+              'caproto-get = caproto.comamndline.get:main',
+              'caproto-put = caproto.comamndline.put:main',
+              'caproto-monitor = caproto.comamndline.monitor:main',
+              'caproto-repeater = caproto.comamndline.repeater:main',
               'caproto-defaultdictserver = caproto.examples.defaultdictserver:main',
               'caproto-spoof-beamline = caproto.examples.spoof_beamline:main',
           ],
